Changes name from LevelManger.java to LevelManagerImpl.java, due to the interface creation and adaption.
